Transfer ManageSuggestions to Suggestions page
